+++ fire_seq_search_server/src/load_notes/mod.rs
@@ -15,7 +15,7 @@ pub fn read_specific_directory(path: &str) -> Vec<(String, String)> {
     }
     // debug!("Note titles: {:?}", &note_filenames);
     let result: Vec<(String,String)> = note_filenames.par_iter()
-        .map(|note|  read_md_file(&note))
+        .map(|note|  read_md_file_and_parse(&note))
         .filter(|x| (&x).is_some())
         .map(|x| x.unwrap())
         .collect();
@@ -41,7 +41,7 @@ pub fn read_specific_directory(path: &str) -> Vec<(String, String)> {
 ///
 /// If input is a directory or DS_STORE, return None
 ///
-pub fn read_md_file(note: &std::fs::DirEntry) -> Option<(String, String)> {
+pub fn read_md_file_and_parse(note: &std::fs::DirEntry) -> Option<(String, String)> {
     if let Ok(file_type) = note.file_type() {
         // Now let's show our entry's file type!
         debug!("{:?}: {:?}", note.path(), file_type);
@@ -77,6 +77,7 @@ pub fn read_md_file(note: &std::fs::DirEntry) -> Option<(String, String)> {
     };
 
 
+    // Now we do some parsing for this file
     let content:String = parse_to_plain_text(&content);
 
     Some((note_title.to_string(),content))
