+++ fire_seq_search_server/tests/unit_test_load_notes.rs
@@ -3,9 +3,35 @@ use fire_seq_search_server::load_notes::read_specific_directory;
 #[test]
 fn load_articles() {
     let r = read_specific_directory("tests/resource/pages");
-    assert_eq!(r.len(), 6);
+    assert_eq!(r.len(), 7);
     for (title,body) in &r{
         assert!(title.len()>0);
         assert!(body.len()>0);
     }
+}
+
+use fire_seq_search_server::markdown_parser::parse_to_plain_text;
+
+
+fn read_file_to_line(relative_path: &str) -> String {
+    let path = vec![String::from("tests/resource/pages"),
+                    relative_path.to_string()];
+    let path = path.join("/");
+    std::fs::read_to_string(&path)
+        .expect("Should have been able to read the file")
+}
+
+
+#[test]
+fn parse() {
+    let md = read_file_to_line("blog_thunderbird_zh.md");
+    let result = parse_to_plain_text(&md);
+    assert!(result.contains("Aug 3, 2021 - 使用 git shallow clone 下载并编译 Thunderbird"));
+    assert!(!result.contains("https://developer.thunderbird.net/thunderbird-development/getting-started"));
+
+}
+
+#[test]
+fn exclude_advance_query() {
+
 }
